Add a fallback mechanism to the validator after a HIBP API failure
Previously, in the event of a HIBP API failure, the password was assumed
to be valid. Instead of allowing a potentially compromised password,
fallback to checking with Django's CommonPasswordValidator. While an
incomplete list, the passwords are generated from the same HIBP
database.

Details on Django's list of common passwords can be seen at:

https://docs.djangoproject.com/en/2.0/topics/auth/passwords/#django.contrib.auth.password_validation.CommonPasswordValidator