Remove handling for malformed 200 response
It is reasonable to expect that, from time to time, HIBP will timeout or
require maintenance and return a non-200 response.

However, given a 200 response, we should be confident that HIBP is
honoring its API and returning a body in the expected, documented
format. Therefore, remove the added complexity to handle such a error
case. Remove all non-representative tests.