Match DEFAULT_HELP_MESSAGE to Django's CommonPasswordValidator
Django recently updated the help text of CommonPasswordValidator. For
details, see the upstream commit:

https://github.com/django/django/commit/42b9a23267f14be39b9b00958e18d5746783208e

Rather than copy and paste the help text every time it is updated,
directly reuse the value from CommonPasswordValidator.

This has the side effect of using a single instance of
CommonPasswordValidator instead of re-instantiating it every time it is
needed.