Add Django master branch to the test matrix
This helps ensure that the project remains compatible with Django as
either evolves. When it gets close to the next Django release,
pwned-passwords-django will be in a more compatible position as it will
have continuously tested against the master branch.

Dropping the upper bound also allows projects which depend on both
Django and pwned-passwords-django to test against the Django master
branch without a version conflict.