Remove type annotations and mypy CI.
This is something of a protest against PEP 561, which requires that
packages take the step of adding an extra file -- "py.typed" -- to
indicate to mypy that they provide type annotations. In theory, the
justification for this was the continued existence of, and packages'
desire to maintain compatibility with, Python 2. But this
justification expired along with Python 2. Requiring maintainers who
supply inline type annotations to take additional steps to notify mypy
of their existence is thus an unjustifiable burden. It becomes more
unjustifiable in light of the fact that projects which produce a
single Python file -- this project is not one, but the author
maintains several which are -- must reorganize themselves and alter
their packaging process to support adding and distributing the extra
"py.typed" file.

As a result, this package will no longer be supporting type
annotations, or checking them via CI. If and when PEP 561 is modified
or superseded to no longer require additional effort from package
maintainters beyond providing inline annotations, this policy may be
revisited.