Reorganize and expand tox/CI config.
This adds a couple packaging-related items, resolves the issue with
the sphinxcontrib-spelling deprecation warning, and picks up the
config changes from PR #33, while better documenting and organizing
the tox config.

This also drops Python 3.5 and adds Python 3.9 in CI, which will be
reflected in later commits and a followup release.