Include `py.typed` so type annotations are exported
Per PEP 561, this file must be included so that the excellent type hints
in this package can be consumed by third parties!

https://peps.python.org/pep-0561/#packaging-type-information

Quick testing locally with `setuptools` and `build` indicates to me that
this file is distributed in the resulting wheel (meaning that the module
can be typechecked), but you should confirm this with your own
distribution processes!

One extra change: Complete `__init__` annotations
=================================================

Also, while we're at it, explicitly annotate the return type of
`__init__`, which is required by PEP 484:

> (Note that the return type of `__init__` ought to be annotated with
> `-> None`. The reason for this is subtle. If `__init__` assumed a
> return annotation of `-> None`, would that mean that an argument-less,
> un-annotated `__init__` method should still be type-checked? Rather
> than leaving this ambiguous or introducing an exception to the
> exception, we simply say that `__init__` ought to have a return
> annotation; the default behavior is thus the same as for other
> methods.)

https://peps.python.org/pep-0484/